+++ JsonToolsNppPlugin/JSONTools/JsonGrepper.cs
@@ -2,8 +2,12 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Threading;
 using System.Net;
+using System.Net.Http;
+using System.Net.Http.Headers;
+using System.Threading;
+using System.Threading.Tasks;
+using JSON_Tools.Utils;
 
 namespace JSON_Tools.JSON_Tools
 {
@@ -20,14 +24,13 @@ namespace JSON_Tools.JSON_Tools
         public JObject exceptions;
         public Dictionary<string, string> fname_strings;
 		public JsonParser json_parser;
-        private static readonly WebClient webClient = new WebClient();
+        private static readonly HttpClient httpClient = new HttpClient();
         public int max_threads_parsing;
-        public int max_api_request_threads;
 
-		public JsonGrepper(JsonParser json_parser = null, int max_threads = 4, int max_api_request_threads = 8)
+		public JsonGrepper(JsonParser json_parser = null, 
+            int max_threads = 4)
 		{
             this.max_threads_parsing = max_threads;
-            this.max_api_request_threads = max_api_request_threads;
             fname_strings = new Dictionary<string, string>();
 			fname_jsons = new JObject();
             exceptions = new JObject();
@@ -40,18 +43,7 @@ namespace JSON_Tools.JSON_Tools
 				this.json_parser = json_parser;
             }
             // security protocol addresses issue: https://learn.microsoft.com/en-us/answers/questions/173758/the-request-was-aborted-could-not-create-ssltls-se.html
-            ServicePointManager.SecurityProtocol = (SecurityProtocolType)3072;
-		}
-
-        private static void InitializeWebClient(WebClient wc)
-        {
-            wc.Headers.Clear();
-            // add a header saying that this client accepts only JSON
-            wc.Headers.Add("Accept", "application/json");
-            // add a user-agent header saying who you are
-            wc.Headers.Add("User-Agent", "JsonTools Notepad++ plugin");
-            // see https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers#request_context
-            // for more info on HTTP request headers
+            ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12;
         }
 
         /// <summary>
@@ -156,35 +148,7 @@ namespace JSON_Tools.JSON_Tools
         }
 
         /// <summary>
-        /// Downloads JSON from APIs. Can use multiple threads.
-        /// </summary>
-        /// <param name="urls"></param>
-        public void GetJsonFromApis(string[] urls)
-        {
-            if (urls.Length == 0)
-                return;
-            if (max_api_request_threads > 1)
-            {
-                GetJsonFromApisMultithreaded(urls);
-                return;
-            }
-            foreach (string url in urls)
-            {
-                InitializeWebClient(webClient);
-                try
-                {
-                    fname_strings[url] = webClient.DownloadString(new Uri(url));
-                }
-                catch (Exception ex)
-                {
-                    exceptions[JObject.FormatAsKey(url)] = new JNode(ex.ToString(), Dtype.STR, 0);
-                }
-            }
-            ParseJsonStringsThreaded();
-        }
-
-        /// <summary>
-        /// Concurrently send API requests to several URLs (divides them among multiple threads)
+        /// Asynchronously send API requests to several URLs
         /// For each URL where the request succeeds, try to parse the JSON returned.
         /// If the JSON returned is valid, add the JSON to fname_jsons.
         /// Populate a HashSet of all URLs for which the request succeeded
@@ -192,53 +156,52 @@ namespace JSON_Tools.JSON_Tools
         /// </summary>
         /// <param name="urls"></param>
         /// <returns></returns>
-        private void GetJsonFromApisMultithreaded(string[] urls)
+        public async Task GetJsonFromApis(string[] urls)
         {
-            List<Thread> threads = new List<Thread>();
-            foreach (object[] urls_this_thread in DivideObjectsBetweenThreads(urls, max_api_request_threads))
+            var json_tasks = new Task[urls.Length];
+            for (int ii = 0; ii < urls.Length; ii++)
             {
                 // if (!fname_jsons.children.ContainsKey(url))
-                // could stop if already downloaded, but it is probably better to allow duplication of labor, 
+                // it is probably better to allow duplication of labor,
                 // so that the user can get new JSON if something changed
-                WebClient newWebClient = new WebClient();
-                Thread thread = new Thread(
-                    () => GetJsonFromApis_Task(
-                        urls_this_thread,
-                        newWebClient,
-                        fname_strings,
-                        exceptions)
-                );
-                threads.Add(thread);
-                thread.Start();
+                json_tasks[ii] = GetJsonStringFromApiAsync(urls[ii]);
             }
-            foreach (Thread thread in threads)
-                thread.Join();
+            await Task.WhenAll(json_tasks);
             // now parse all the JSON strings that were downloaded
             ParseJsonStringsThreaded();
         }
 
         /// <summary>
-        /// the task of a single thread in GetJsonFromApisMultithreaded
-        /// </summary>
-        /// <param name="urls_this_thread"></param>
-        private static void GetJsonFromApis_Task(object[] urls_this_thread, 
-                                          WebClient webClient,
-                                          Dictionary<string, string> fname_strings,
-                                          JObject exceptions)
+		/// Send an asynchronous request for JSON to an API.
+		/// If the request succeeds, add the JSON string to fname_strings
+		/// If the request raises an exception, add the exception string to exceptions
+		/// </summary>
+		/// <param name="url"></param>
+		/// <returns></returns>
+		public async Task GetJsonStringFromApiAsync(string url)
         {
-            foreach (object urlobj in urls_this_thread)
+            InitializeHttpClient(httpClient);
+            string formatted_url = JObject.FormatAsKey(url);
+            try
             {
-                string url = (string)urlobj;
-                InitializeWebClient(webClient);
-                try
-                {
-                    fname_strings[url] = webClient.DownloadString(url);
-                }
-                catch (Exception ex)
-                {
-                    exceptions[JObject.FormatAsKey(url)] = new JNode(ex.ToString(), Dtype.STR, 0);
-                }
+                Task<string> stringTask = httpClient.GetStringAsync(url);
+                fname_strings[formatted_url] = await stringTask;
             }
+            catch (Exception ex)
+            {
+                exceptions[formatted_url] = new JNode(ex.ToString(), Dtype.STR, 0);
+            }
+        }
+
+        private static void InitializeHttpClient(HttpClient hc)
+        {
+            hc.DefaultRequestHeaders.Clear();
+            // add a header saying that this client accepts only JSON
+            hc.DefaultRequestHeaders.Add("Accept", "application/json");
+            // add a user-agent header saying who you are
+            hc.DefaultRequestHeaders.Add("User-Agent", $"JsonTools Notepad++ plugin v{Npp.AssemblyVersionString()}");
+            // see https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers#request_context
+            // for more info on HTTP request headers
         }
 
         /// <summary>
@@ -252,14 +215,13 @@ namespace JSON_Tools.JSON_Tools
 			//fname_lints.Clear();
         }
 
-
         /// <summary>
         /// figure out how to divide up task_count tasks equally among num_threads threads
         /// </summary>
         /// <param name="task_count"></param>
         /// <param name="num_threads"></param>
         /// <returns></returns>
-        private IEnumerable<int> TaskCountPerThread(int task_count, int num_threads)
+        private static IEnumerable<int> TaskCountPerThread(int task_count, int num_threads)
         {
             int start = 0;
             int things_per_thread = task_count / num_threads;
@@ -285,7 +247,7 @@ namespace JSON_Tools.JSON_Tools
         /// <param name="objs"></param>
         /// <param name="num_threads"></param>
         /// <returns></returns>
-        private IEnumerable<object[]> DivideObjectsBetweenThreads(object[] objs, int num_threads)
+        private static IEnumerable<object[]> DivideObjectsBetweenThreads(object[] objs, int num_threads)
         {
             int start = 0;
             foreach (int count in TaskCountPerThread(objs.Length, num_threads))