+++ JsonToolsNppPlugin/Utils/Settings.cs
@@ -89,10 +89,6 @@ namespace JSON_Tools.Utils
         [Description("How many threads to use for parsing JSON files obtained by JsonGrep and API requester"),
             Category("Grep and API requests"), DefaultValue(4)]
         public int max_threads_parsing { get; set; }
-
-        [Description("How many threads to use for requesting JSON from remote APIs?"),
-            Category("Grep and API requests"), DefaultValue(8)]
-        public int max_api_request_threads { get; set; }
         #endregion
 
         #region STYLING